feat(Server): Adds re-export of AtomicState from module with the 'use client' directive
