+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "atomic-state",
-  "version": "2.2.8",
+  "version": "2.3.0",
   "description": "Atomic State is a managment library for React",
   "main": "dist/index.js",
   "repository": {
