--- src/server/index.ts
@@ -1,2 +1,0 @@
-"use client"
-export { AtomicState } from "../"
