+++ src/server/index.tsx
@@ -1,0 +1,2 @@
+"use client"
+export { AtomicState } from "../mod"
